Refactor get_data_files() function in setup.py.
The ``get_data_files()`` function now takes a boolean ``filesonly`` parameter,
if True, it returns a list suitable for the ``package_data`` parameter of
``setuptools.setup()``, otherwise it returns a list which is compatible with
``distutils.command.install_data.install_data.data_files``.

We can now remove distutils entirely from setup.py, but I decided to make this
function compatible with both distutils and setuptools because Python
packaging has been in turmoil for a while now. (Hopefully, it's done, but I'm
not holding my breath.)

 * FIXES a problem with some .mo files not getting installed (if they didn't
   already have a directory for their LOCALE in build/).

 * FIXES a bug where all .po files were being installed, even though they were
   added to the ``exclude_package_data`` parameter of ``setuptools.setup()``.

 * FIXES (along with commit ba20ca82f65be29cecab88cbd305a31bdb928108) a bug
   where the robots.txt file was not being packaged correctly.