The identity is based on the public key of the router.
Specifically it is the SHA-1 hash of the DER encoding of an ASN.1
RSA public key.

(cherry picked from commit 6b57521a522abaa8f5fdd158708f382293a59e48)
Signed-off-by: Isis Lovecruft <isis@torproject.org>

Conflicts:
	scripts/gen_bridge_descriptors

Matt and I both fixed the same bug, but it turns out we were both a tiny bit
wrong, I believe, in different ways: I was improperly PEM-encoding the OR
keys, and wasn't using ASN.1 format. Matt was using a dump of SIDPKey as the
OR bridge's public identity key -- PyOpenSSL has this rather odd API where you
have to dump the public key from the public cert to access it. Also, it's
necessary to base64-encode the digest of the identity key, and strip the '='
character base64 padding.

We *might* still be doing it wrong and missing the DER-encoding step.