Cleanup logging of memory addresses in prepopulateRings().
The logging before was quite messy.

The sub hashrings are identified by the HMAC of a unique string. It
turns out that the "unique string" which is used in this case is
actually something which looks like the class.__repr__() for a
bridgedb.Bridges.BridgeHolder subclass. This is *probably* unique
because it contains memory addresses of the newly generated hmac_fn()s,
which are likely (though not necessarily) unique. This is a bit gross,
in my opinion, but it's not really a problem.

However, the log messages this function generated are a problem. They
contain memory addresses, which isn't the greatest idea, ignoring all
other reasons, simply because it's illegible.

 * REFACTOR logging statements in
   bridgedb.Bridges.FilteredBridgeSplitter.prepopulateRings() to remove
   logging of hmac_fn() memory addresses.
 * ADD a long comment detailing the "unique strings" which identify a
   sub hashring, and how they should probably be changed someday.