Add generic Dist.Distributor class for common functionality.
This class currently serves very little purpose because I am still
nervous about making major changes to the IPBasedDistributor and
EmailBasedDistributor classes. However, I did add functionality for
naming a Distributor's hashrings after the distributor's class, so that
which hashring a bridge is being added to is much clearer in the logs.

 * ADD bridgedb.Dist.Distributor class.
 * CHANGE IPBasedDistributor and EmailBasedDistributor to inherit from
   Distributor.
 * ADD functionality for naming hashrings after their distributor.