Always expect a Deferred from ReCaptcha.checkSolution().
Currently, we `checkSolution()` will return a RecaptchaResponse if the
HTTP POST argument fields are blank. Otherwise, it returns a
deferred. This is bad. It means we have to do all kinds of checks to see
if we can add callbacks on the returned object. It should always return
a deferred which callbacks with a RecaptchaResponse, and so we should
test for that and fail if that is not what is happening.