Rewrite HTML templates to pass template namespace parameters.
In order to use translated strings from the bridgedb.strings module, we
need to pass the namespace parameters (i.e., the arguments in
HTTPServer.py to `template.render()`) from the parent template
(base.html) to the inheriting templates.

While we're at it, we should pass information on the requested lang, so
that translations don't suddenly turn off if you go to a different page,
and info on whether the page should be rendered RTL. This fixes a slight
bug where the pages deriving from
`bridgedb.HTTPServer.CaptchaProtectedResource` weren't being correctly
translated or RTL-rendered when they should be.

And also while we're at it, we should probably not do *anything* inside
of `bridgedb.HTTPServer.*Resource.render_*` methods outside of the
try/except block, because any error will be rendered as a traceback to
the client.

Additionally, we should set the content type and encoding in the headers
of every response before returning it.