Make publicKey and secretKey attributes of the ICaptcha interface.
Both the reCaptcha and GimpCaptcha subclasses of the generic base
implementation, bridgedb.captcha.Captcha, of the ICaptcha interface have
publicKey and secretKey attributes. In fact, it would be hard to
generate a secret CAPTCHA challenge_string which was retrievable later
if there weren't any keypairs involved. Ergo, the keypair should be part
of the interface specification.