Add TODO: bridgedb.Bridges.BridgeBlock should be a zope.interface.Interface.
This does not count as a base class. There is absolutely no point to
it. I haven't bothered to do a git blame to see whose idea this was; it
doesn't matter. This is NOT A BASE CLASS. Especially when it is an
old-style class without a proper method resolution ordering algorithm.