Constify the unrealistic maximum number of bridges per bucket
We will never, ever have 1 million bridges, but because of this we
will also never be able to satisfy the requirement of giving a bucket
one million bridges. So, when we try distributing the bridges evenly
between buckets, the pseudo-distributors that are defined as needing
'*' bridges will receive as many as are available to be allocated to
them.