Add obfs4 to the HTTPS Distributor's PT options dropdown.
This is temporary. It is needed in order to make sure that the
regression tests for #12932 are functioning correctly. Because #12932 is
about the way PT arguments are presented to the user in the bridge
lines, to test the fixes for it, we must go through the UI (as the tests
in `lib/bridgedb/test/test_https.py` do).