Move Bridge backwards compat code into new base classes.
This changes the behaviour, such that if BridgeBackwardsCompatibility is
inherited from, then the compatibility methods and attributes are always
available, no matter how the class was instantiated (i.e. the old
method, Bridge.getID() is always available). However, the assertions and
errors raised in the BridgeBackwardsCompatiblity._backwardsCompatible
method are only run *if* the inherited class is instantiated *with*
arguments (as was the legacy manner of doing things).

 * ADD bridges.BridgeBase and bridges.BridgeBackwardsCompatibility
   classes.