Add qrcode==5.0.1 and pillow>=2.6.1 to Pip requirements.txt file.
Pillow, in Debian jessie, is in the `python-imaging` package (currently
version 2.6.1).  However, on PyPI, Pillow is simply `pillow` (currently
version 2.7.0). Both of these are backwards compatible with the
python-imaging package in Debian wheezy (currently version 1.1.6, and
likely to never get updated, because it is being replaced with pillow).
And any of these will currently work with BridgeDB.