Fix error for initializeGnuPG() return value when signing fails.
When the test signing in bridgedb.crypto.initializeGnuPG() fails, the
return value would be `None`, rather than the gpg interface and the
signing function, `(gpg, gpgSignMessage)`, or `(None, None)` as it does
with the other failure modes.  This would cause the EmailServerContext
to raise a TypeError because `None` is not iterable, and cannot be
assigned to two variables.

We now return `(None, None)` when the test signing fails.

This is a bugfix on #10385. [0]

[0]: https://bugs.torproject.org/10385