Revert back to using Bridge.getID() in filterAssignBridgesToRing().
Bridge.getID() returns the Bridge's identity key digest in binary, and
Bridge.fingerprint return the digest in hexadecimal.

While this is not a bug, we should avoid randomly changing the encoding
used to assign Bridges to hashrings, as it might result in e.g. a Bridge
which was previously assigned to hashing 0 based on the binary digest
getting a duplicate assignment in hashring 1 based on the
hexadecimal-encoded digest.

 * REVERT change made in 8574c134 to the encoding of Bridge identity key
   digest used for assigning a Bridge to a hashring.