Check envvar before using easy_install to install setup.py requires.
If there is an environment variable BRIDGEDB_INSTALL_DEPENDENCIES=0,
then the setup.py script will not use easy_install to install
dependencies.  The default when using `make install` is to avoid using
easy_install; the default when calling `python setup.py install` is to
use easy_install.

Until now, on BridgeDB's production server, `echo > requirements.txt`
was run to avoid allowing easy_install to search for dependencies.  This
removes the need to do that.  We shouldn't be erasing the
requirements.txt file anymore, because it is kept under revision control
and versioneer (which uses `git describe --tags --dirty --always` to
dynamically determine the BridgeDB package version at install time) will
add '-dirty' to the end of the version number if we change anything in
the repo.  Adding '-dirty' to the package version causes it to be
invalid according to PEP440, causing other random failures such as the
inability to discover scripts and packaged data which pkg_resources.